pyproject.toml
[tool.poetry]
name = "vkbottle-example"
version = "0.1.0"
description = "Getting started with vkbottle"
authors = ["AnyGlitch <WhyGlitches@gmail.com>"]
keywords = [
  "vk",
  "vk-api",
  "vkbottle",
  "vkbottle-example",
  "longpoll-api",
  "callback-api",
]

[tool.poetry.dependencies]
python = "^3.10.0"
vkbottle = "^4.2.1"
aerich = "^0.6.2"
tortoise-orm = "^0.19.0"
asyncpg = "^0.25.0"
python-decouple = "^3.6.0"

[tool.poetry.dev-dependencies]
loguru = "^0.6.0"
pre-commit = "^2.17.0"
pytest = "^7.1.1"
mypy = "^0.942.0"
black = "^22.1.0"
isort = "^5.10.1"

[tool.aerich]
tortoise_orm = "source.config.ORM"
location = "./migrations"
src_folder = "./."

[tool.mypy]
ignore_missing_imports = true

[tool.black]
line-length = 80

[tool.isort]
line_length = 80
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
